Reorganise code into standard Python package layout
Move the CLI into `datascrubber/cli.py` and add a module entry point in
`__main__.py` so that the tool can be invoked with `python -m
datascrubber`.
