+++ bin/datascrubber
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+import sys
+import datascrubber.cli
+
+
+def main():
+    return datascrubber.cli.main()
+
+
+if __name__ == '__main__':
+    sys.exit(main())
