+++ datascrubber/__main__.py
@@ -1,0 +1,6 @@
+import sys
+from datascrubber.cli import main
+
+if __name__ == "__main__":
+    sys.exit(main())
+
