+++ datascrubber/sql/scrub_publishing_api.sql
